+++ app/backend/UserSession.h
@@ -9,7 +9,7 @@ class UserSession : public QObject {
     Q_PROPERTY(QString username READ username WRITE setUsername NOTIFY usernameChanged)
     Q_PROPERTY(QString token READ token WRITE setToken NOTIFY tokenChanged)
     Q_PROPERTY(qint64  userId READ userId WRITE setUserId NOTIFY userIdChanged)
-    Q_PROPERTY(int balance READ balance WRITE setBalance NOTIFY balanceChanged)
+    Q_PROPERTY(double balance READ balance WRITE setBalance NOTIFY balanceChanged)
     Q_PROPERTY(QString notice READ notice WRITE setNotice NOTIFY noticeChanged)
 
 public:
@@ -27,8 +27,8 @@ public:
     Q_INVOKABLE void saveToSettings();       // 登录后保存 token + username
     qint64  userId() const { return m_userId; }
     void    setUserId(qint64 id);
-    int     balance() const { return m_balance; }
-    void    setBalance(int balance);
+    double  balance() const { return m_balance; }
+    void    setBalance(double balance);
     QString notice() const { return m_notice; }
     void    setNotice(const QString& notice);
 
@@ -44,7 +44,7 @@ private:
     QString m_username;
     QString m_token;
     qint64  m_userId = 0;
-    int     m_balance = 0;
+    double  m_balance = 0.0;
     QString m_notice;
 };
